// src/app/shared/interfaces/Film/film.ts
export interface Film {
  characterIds: number[];
  characters: string[];
  director: string;
  episode_id: string;
  id: number;
  opening_crawl: string;
  planetIds: number[];
  planets: string[];
  producer: string;
  title: string;
  release_date: string;
  speciesIds: number[];
  species: string[];
  starshipIds: number[];
  starships: string[];
  url: string;
  vehicleIds: number[];
  vehicles: string[];
}